[BOD-271] Move LEGO components to dedicated folders
