+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "agroptima-design-system",
-  "version": "0.13.1",
+  "version": "0.14.0",
   "scripts": {
     "dev": "npm run storybook",
     "storybook": "storybook dev -p 6006 --ci",
