+++ src/atoms/Card/Card.scss
@@ -1,7 +1,7 @@
-@use '../settings/color_alias';
-@use '../settings/typography/content' as typography;
-@use '../settings/config';
-@use '../settings/breakpoints';
+@use '../../settings/color_alias';
+@use '../../settings/typography/content' as typography;
+@use '../../settings/config';
+@use '../../settings/breakpoints';
 
 .card {
   display: flex;
