+++ src/atoms/Card/Card.tsx
@@ -1,6 +1,6 @@
 import './Card.scss'
 import React from 'react'
-import { classNames } from '../utils/classNames'
+import { classNames } from '../../utils/classNames'
 
 export type Variant = 'primary'
