+++ src/atoms/Card/index.ts
@@ -1,0 +1,6 @@
+import { Card } from './Card'
+import { CardHeader } from './CardHeader'
+import { CardContent } from './CardContent'
+import { CardFooter } from './CardFooter'
+
+export { Card, CardHeader, CardContent, CardFooter }
