+++ src/atoms/CardsTable/CardsTable.tsx
@@ -1,4 +1,4 @@
-import { classNames } from '../utils/classNames'
+import { classNames } from '../../utils/classNames'
 import './CardsTable.scss'
 
 export type Variant = 'primary'
