+++ src/atoms/CardsTable/CardsTableHeader.tsx
@@ -1,4 +1,4 @@
-import { classNames } from '../utils/classNames'
+import { classNames } from '../../utils/classNames'
 import { Alignment } from './CardsTableCell'
 
 export interface CardsTableHeaderProps
