+++ src/atoms/CardsTable/CardsTableRow.tsx
