+++ src/atoms/CardsTable/index.ts
@@ -1,0 +1,15 @@
+import { CardsTable } from './CardsTable'
+import { CardsTableHead } from './CardsTableHead'
+import { CardsTableHeader } from './CardsTableHeader'
+import { CardsTableRow } from './CardsTableRow'
+import { CardsTableBody } from './CardsTableBody'
+import { CardsTableCell } from './CardsTableCell'
+
+export {
+  CardsTable,
+  CardsTableHead,
+  CardsTableHeader,
+  CardsTableRow,
+  CardsTableBody,
+  CardsTableCell,
+}
