+++ src/atoms/Menu/Menu.scss
@@ -1,7 +1,7 @@
-@use '../settings/color_alias';
-@use '../settings/typography/content' as typography;
-@use '../settings/config';
-@use '../settings/depth';
+@use '../../settings/color_alias';
+@use '../../settings/typography/content' as typography;
+@use '../../settings/config';
+@use '../../settings/depth';
 
 .menu {
   list-style-type: none;
