+++ src/atoms/Menu/MenuOption.tsx
@@ -1,7 +1,7 @@
 import './Menu.scss'
 import React from 'react'
-import { Icon, IconType } from './Icon'
-import { classNames } from '../utils/classNames'
+import { Icon, IconType } from '../Icon'
+import { classNames } from '../../utils/classNames'
 
 export type Variant = 'primary'
