+++ src/atoms/Menu/index.ts
@@ -1,0 +1,4 @@
+import { Menu } from './Menu'
+import { MenuOption } from './MenuOption'
+
+export { Menu, MenuOption }
