+++ src/stories/CardsTable.stories.js
@@ -1,11 +1,6 @@
 import React from 'react'
 
-import { CardsTable } from '../atoms/CardsTable'
-import { CardsTableHead } from '../atoms/CardsTableHead'
-import { CardsTableHeader } from '../atoms/CardsTableHeader'
-import { CardsTableRow } from '../atoms/CardsTableRow'
-import { CardsTableBody } from '../atoms/CardsTableBody'
-import { CardsTableCell } from '../atoms/CardsTableCell'
+import { CardsTable, CardsTableHead, CardsTableHeader, CardsTableRow, CardsTableBody, CardsTableCell } from '../atoms/CardsTable'
 import { IconButton } from '../atoms/IconButton'
 import { Badge } from '../atoms/Badge'
