+++ src/stories/Changelog.stories.mdx
@@ -3,6 +3,8 @@ import { Meta } from "@storybook/addon-docs";
 <Meta title="Changelog" />
 # Changelog
 
+## 0.14.0
+- Added Menu component to Storybook.
 
 ## 0.13.0
 - Added CardMenuOption component to Storybook.
